+++ cookiecutter-django-ida/{{cookiecutter.repo_name}}/{{cookiecutter.project_name}}/settings/utils.py
@@ -17,14 +17,22 @@ def get_env_setting(setting):
 
 
 def get_logger_config(
-    logging_env="no_env",
-    debug=False,
-    service_variant='{{cookiecutter.repo_name}}',
+    logging_env: str = "no_env",
+    debug: bool = False,
+    service_variant: str = 'enterprise-access',
+    format_string: str = None,
 ):
     """
-    Return the appropriate logging config dictionary.
+    Return the appropriate logging config dictionary, to be assigned to the LOGGING var in settings.
 
-    You should assign the result of this to the LOGGING var in your settings.
+    Arguments:
+        logging_env (str): Environment name.
+        debug (bool): Debug logging enabled.
+        service_variant (str): Name of the service.
+        format_string (str): Override format string for your logfiles.
+
+    Returns:
+        dict(string): Returns a dictionary of config values
     """
     hostname = platform.node().split(".")[0]
     syslog_format = (
@@ -39,13 +47,18 @@ def get_logger_config(
 
     handlers = ['console']
 
+    standard_format = format_string or (
+        '%(asctime)s %(levelname)s %(process)d [%(name)s] '
+        '[user %(userid)s] [ip %(remoteip)s] '
+        '%(filename)s:%(lineno)d - %(message)s'
+    )
+
     logger_config = {
         'version': 1,
         'disable_existing_loggers': False,
         'formatters': {
             'standard': {
-                'format': '%(asctime)s %(levelname)s %(process)d '
-                          '[%(name)s] [user %(userid)s] [ip %(remoteip)s] %(filename)s:%(lineno)d - %(message)s',
+                'format': standard_format,
             },
             'syslog_format': {'format': syslog_format},
             'raw': {'format': '%(message)s'},